requirements.txt
playwright
requests
pandas
beautifulsoup4
python-dotenv
apscheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot
openpyxl
sqlalchemy
